pyproject.toml
[project]
name = "applypilot"
version = "0.3.0"
description = "AI-powered end-to-end job application pipeline"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
authors = [{name = "Pickle-Pixel"}]
keywords = ["job-search", "automation", "ai", "resume", "apply", "job-application"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "httpx>=0.24",
    "beautifulsoup4>=4.12",
    "playwright>=1.40",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "pypdf",
    "reportlab",
    "openai>=1.40",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "supabase>=2.0",
    ]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]
applypilot = "applypilot.cli:app"

[project.urls]
Homepage = "https://github.com/Pickle-Pixel/ApplyPilot"
Repository = "https://github.com/Pickle-Pixel/ApplyPilot"
Issues = "https://github.com/Pickle-Pixel/ApplyPilot/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/applypilot"]

[tool.hatch.build]
artifacts = ["src/applypilot/config/*.yaml"]

[tool.ruff]
target-version = "py311"
line-length = 120
